+++ core/tw-sched.c
@@ -250,7 +250,7 @@ static void tw_sched_batch(tw_pe * me) {
         if(g_st_real_time_samp && tw_clock_read() - g_st_real_samp_start_cycles > g_st_real_time_samp)
         {
             tw_clock current_rt = tw_clock_read();
-            st_collect_data(me, current_rt / g_tw_clock_rate);
+            st_collect_data(me, (double)current_rt / g_tw_clock_rate);
             g_st_real_samp_start_cycles = tw_clock_read();
         }
     }
@@ -370,7 +370,7 @@ static void tw_sched_batch_realtime(tw_pe * me) {
         if(g_st_real_time_samp && tw_clock_read() - g_st_real_samp_start_cycles > g_st_real_time_samp)
         {
             tw_clock current_rt = tw_clock_read();
-            st_collect_data(me, current_rt / g_tw_clock_rate);
+            st_collect_data(me, (double)current_rt / g_tw_clock_rate);
             g_st_real_samp_start_cycles = tw_clock_read();
         }
     }
@@ -562,7 +562,7 @@ void tw_scheduler_conservative(tw_pe * me) {
             if(g_st_real_time_samp && tw_clock_read() - g_st_real_samp_start_cycles > g_st_real_time_samp)
             {
                 tw_clock current_rt = tw_clock_read();
-                st_collect_data(me, current_rt / g_tw_clock_rate);
+                st_collect_data(me, (double)current_rt / g_tw_clock_rate);
                 g_st_real_samp_start_cycles = tw_clock_read();
             }
         }
@@ -585,7 +585,7 @@ void tw_scheduler_conservative(tw_pe * me) {
     if (g_st_real_time_samp)
     {
         // collect data one final time to account for time between last sample and sim end time
-        st_collect_data(me, tw_clock_read() / g_tw_clock_rate);
+        st_collect_data(me, (double)tw_clock_read() / g_tw_clock_rate);
         st_buffer_finalize(g_st_buffer_rt, RT_COL);
     }
     if (g_st_ev_trace)
@@ -640,7 +640,7 @@ void tw_scheduler_optimistic(tw_pe * me) {
     if (g_st_real_time_samp)
     {
         // collect data one final time to account for time between last sample and sim end time
-        st_collect_data(me, tw_clock_read() / g_tw_clock_rate);
+        st_collect_data(me, (double)tw_clock_read() / g_tw_clock_rate);
         st_buffer_finalize(g_st_buffer_rt, RT_COL);
     }
     if (g_st_ev_trace)
@@ -700,7 +700,7 @@ void tw_scheduler_optimistic_realtime(tw_pe * me) {
     if (g_st_real_time_samp)
     {
         // collect data one final time to account for time between last sample and sim end time
-        st_collect_data(me, tw_clock_read() / g_tw_clock_rate);
+        st_collect_data(me, (double)tw_clock_read() / g_tw_clock_rate);
         st_buffer_finalize(g_st_buffer_rt, RT_COL);
     }
     if (g_st_ev_trace)
